recommited Kick confirmation modal feature

+++ src/app/ui/team-page/team-members/team-members.component.ts
@@ -5,6 +5,9 @@ import { map } from 'rxjs/operators';
 import { TeamsService } from 'src/app/teams/teams.service';
 import { PlayerPointsService } from '../../player-quest/player-points.service';
 
+import { MatDialog } from '@angular/material';
+import { KickDialogComponent } from '../kick-dialog/kick-dialog.component';
+
 @Component({
   selector: 'team-members',
   templateUrl: './team-members.component.html',
@@ -22,6 +25,7 @@ export class TeamMembersComponent implements OnInit {
   readonly memberColumns = ['displayName', 'exp', 'seasonRank', 'kickButton'];
 
   constructor(
+    private dialog: MatDialog,
     private teamsService: TeamsService,
     private playerPointsService: PlayerPointsService
   ) {}
@@ -48,8 +52,16 @@ export class TeamMembersComponent implements OnInit {
     return this.currentUser.membership.isLead && (uid !== this.currentUser.uid);
   }
 
-  removeTeamMember(uid) {
-    this.teamsService.removeTeamMember(uid);
+  removeTeamMember(uid: string, displayName: string) {
+    const kickMemberDialog = this.dialog.open(KickDialogComponent, {
+      data: { displayName }
+    });
+
+    kickMemberDialog.afterClosed().subscribe(result => {
+      if (result) {
+        this.teamsService.removeTeamMember(uid);
+      }
+    });
   }
 
   private getMemberInfo(member: Membership): Observable<Membership> {
